Move test setup folder to be with the other test folder
