+++ src/modules/files/tests/file.e2e.test.ts
@@ -6,7 +6,7 @@ import { NestExpressApplication } from '@nestjs/platform-express'
 import type { File } from '../entities/file.entity.js'
 import { TestContext } from '../../../../test/utils/test-context.js'
 import type { TestUser } from '../../users/tests/setup-user.type.js'
-import { setupTest } from '../../../utils/test-setup/setup.js'
+import { setupTest } from '../../../../test/setup/test-setup.js'
 import { CreateFileDtoBuilder } from './builders/create-file-dto.builder.js'
 import { FileSeeder } from './seeders/file.seeder.js'
 import { FileBuilder } from './builders/file-link.builder.js'
