+++ src/modules/roles/tests/role.e2e.test.ts
@@ -10,7 +10,7 @@ import { Permission } from '../../permissions/permission.enum.js'
 import { UserSeeder } from '../../users/tests/user.seeder.js'
 import { UserEntityBuilder } from '../../users/tests/user-entity.builder.js'
 import { TokenSeeder } from '../../auth/tests/seeders/token.seeder.js'
-import { setupTest } from '../../../utils/test-setup/setup.js'
+import { setupTest } from '../../../../test/setup/test-setup.js'
 import { ClientSeeder } from '../../auth/tests/seeders/client.seeder.js'
 import { TestContext } from '../../../../test/utils/test-context.js'
 import type { TestUser } from '../../users/tests/setup-user.type.js'
