+++ src/utils/validators/tests/is-same-or-after-date.test.ts
@@ -3,7 +3,7 @@ import dayjs from 'dayjs'
 import { validate } from 'class-validator'
 import { expect } from 'expect'
 import { NestExpressApplication } from '@nestjs/platform-express'
-import { setupTest } from '../../test-setup/setup.js'
+import { setupTest } from '../../../../test/setup/test-setup.js'
 import { IsSameOrAfterDateString } from '../is-same-or-after-date.js'
 
 class TestClass {
