+++ src/utils/validators/tests/is-valid-time.test.ts
@@ -2,7 +2,7 @@ import { after, before, describe, it } from 'node:test'
 import { validate } from 'class-validator'
 import { expect } from 'expect'
 import { NestExpressApplication } from '@nestjs/platform-express'
-import { setupTest } from '../../test-setup/setup.js'
+import { setupTest } from '../../../../test/setup/test-setup.js'
 import { IsValidTimeString } from '../is-valid-time.js'
 
 class TestClass {
