+++ test/setup/compile-test-module.ts
@@ -1,5 +1,5 @@
 import { Test, type TestingModule } from '@nestjs/testing'
-import { AppModule } from '../../app.module.js'
+import { AppModule } from '../../src/app.module.js'
 
 export async function compileTestModule (): Promise<TestingModule> {
   const imports = [AppModule.forRoot()]
