+++ test/setup/test-setup.ts
@@ -5,12 +5,12 @@ import { HttpAdapterHost } from '@nestjs/core'
 import type { TestingModule } from '@nestjs/testing'
 import { expect } from 'expect'
 import { NestExpressApplication } from '@nestjs/platform-express'
-import { HttpExceptionFilter } from '../exceptions/http-exception.filter.js'
-import { uuid } from '../../../test/expect/expectUuid.js'
-import { toHaveErrorCode } from '../../../test/expect/expectErrorCode.js'
-import { toHaveStatus } from '../../../test/expect/expectStatus.js'
-import { isEnumValue } from '../../../test/expect/expectEnum.js'
-import { S3Service } from '../../modules/files/services/s3.service.js'
+import { HttpExceptionFilter } from '../../src/utils/exceptions/http-exception.filter.js'
+import { uuid } from '../expect/expectUuid.js'
+import { toHaveErrorCode } from '../expect/expectErrorCode.js'
+import { toHaveStatus } from '../expect/expectStatus.js'
+import { isEnumValue } from '../expect/expectEnum.js'
+import { S3Service } from '../../src/modules/files/services/s3.service.js'
 import { compileTestModule } from './compile-test-module.js'
 
 export interface TestSetup {
